In the prior version of the code, in both xm_wpxp_lhs and xm_wpxp_rhs, there was a single loop (from k = 2 to k = gr%nnzp-1) in which the matrix for both wpxp and xm was set.  With this change, the single loop is being split into two loops -- one for setting the xm portion of the matrix and one for setting the wpxp portion of the matrix -- in both xm_wpxp_lhs and xm_wpxp_rhs.  For now, both loops remain between k = 2 and k = gr%nnzp-1.
Why do this, do you ask?  Well, it has come to my attention that the code
for xm is not set up right.  The code for wpxp is set up right -- in that
wpxp should be specified at both level k = 1 (to wpxp_sfc) and at level
k = gr%nnzp (to 0).  Thus, only levels k = 2 through k = gr%nnzp-1 need to
be looped over.  However, xm should not follow this same pattern, as it
has for years and years.  The xm code is correct at the lowest level, k = 1,
because the k = 1 thermodynamic level is below the model surface.  Thus, it
can simply maintain it's own value during the solving portion, and then be
reset to the value of xm(2) after the fact.  It doesn't effect anything.
However, the current code fixes the value of xm(gr%nnzp) over the entire
run, even though the gr%nnzp thermodynamic level is below the gr%nnzp
momentum level, and is "in play".  Unlike xm(1), xm(gr%nnzp) directly
effects the value of xm(gr%nnzp-1) and the value of wpxp(gr%nnzp-1).

Anyway, some cases have forcing values that should be applied at the top
grid level (ahem ... MPACE-A), but are not being applied at that level, even
though they are being applied at other levels (like level gr%nnzp-1).  This
could conceivably cause a problem, as an unnatural slope between xm(gr%nnzp)
and xm(gr%nnzp-1) could begin to develop.  Regardless, this bug needs to be
fixed.

In the future, the xm loop will loop from k = 2 to k = gr%nnzp.

Baby steps ...

Anyway, the current code change produces the expected results, which are
exact binary matches to all GrADS files for all cases from the previous
version of the code.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@3978 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 46de04e3929e9724b157bf5c47411176bbaf70e7