I added a new input variable, sfc_elevation, to the code.  It is the altitude of the surface (ground) in meters above mean sea level (AMSL).  For CLUBB's purposes, it is read in through the namelist, and then passed into advance_clubb_core.  For host model purposes, it is passed into advance_clubb_core from every grid column.  For all current CLUBB cases, I have simply set it equal to 0.  The results is an exact GrADS binary match for all cases and files.
This has been implemented in order to correct a problem in WRF-CLUBB.  CLUBB
previously defined the surface to be at 0.0 meters.  Any value of gr%zm(1) that
was not 0.0 meters (for example, altocumulus cases), would not call the sfc_var
subroutine, but rather just set them to 0.  However, WRF uses a terrain-following
height coordinate, and sets the lowest model momentum level, equivalent to CLUBB's
gr%zm(1), to the height of the terrain that the column is over.  Thus, gr%zm(1)
basically would only equal 0.0 over the ocean.  Over land, sfc_var was not being
called.  This corrects the problem, as from WRF (or any host model), sfc_elevation
can be set to gr%zm(1), and then passed into advance_clubb_core.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4116 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3030eb363f787f2ce7e4b3564e7c643ff60ef72a