I made a slight adjustment to the call to pdf_closure on momentum levels (which is only called if l_call_pdf_closure_twice is .true.).
Basically, a little-known fact about CLUBB is the setting of p_in_Pa(1)
and exner(1).  Now, p_in_Pa and exner are both thermodynamic-level variables,
and thermodynamic level 1 is below the surface (or below the model lower
boundary).  However, CLUBB does not store the surface pressure, other than
in namelist variable psfc.  Since thermodynamic level 1 is a "junk" level,
where the value of variables never come into play, the values of p_in_Pa(1),
exner(1), and rho(1) are actually set to be the surface pressure, exner,
and density, respectively.  Thus:

p_in_Pa(1) = pressure at momentum level 1 = psfc,
exner(1) = exner at momentum level 1 = (psfc/p0)**kappa, and
rho(1) = density at momentum level 1 = rho_zm(1).

Otherwise, the remainder of the p_in_Pa, exner, and rho arrays are all at
thermodynamic levels.

For more details, look at subroutine hydrostatic.

Most likely, the code is set this way as a left-over from the days when a
second call to pdf_closure was made -- but only at momentum level 1.  The
idea was to find a better way of obtaining wp4 at the surface, which was 
used directly, at that time, in the wp3 equation.  This must have been a
relic from the "pre-ansatz" days.

Anyway, I like the idea of storing the actual surface pressure and exner
at level 1 because (1) that level is junk anyway, and (2) surface pressure
can be a very important variable, especially if CLUBB is being used in a
host model.

However, a slight adjustment is needed for the new second call to pdf_closure
on momentum levels because a standard interpolation of p_in_Pa and exner
won't work for momentum level 1.  Therefore, two new local variables, p_in_Pa_zm
and exner_zm, have been created.  Now, p_in_Pa_zm is an interpolation of p_in_Pa
at all levels except level 1, where p_in_Pa_zm(1) = p_in_Pa(1).  I have set
exner_zm to be directly based on p_in_Pa_zm.

This will effect the code if both the trapezoidal rule and calling pdf_closure
twice are turned on.  When just l_call_pdf_closure_twice is on, the values of
the second-order pdf_closure outputs are effected at level 1.  However, those
variables from that level are never used in the rest of the code.

The differences in pressure and exner between the previous version of the code,
with the interpolation at level 1, and the current version are not a lot with
normal grid spacing.  However, that difference becomes larger (and thus more
significant) when coarser grid spacings are used.

I also noticed that the newer code that allows a sounding to be entered based
on pressure levels did not completely match the code from subroutine hydrostatic.
It did set exner(1) to the actually surface exner.  However, it did not insure
that the value of p_in_Pa(1) was the actually surface pressure -- rather it was
left as the pressure at the lowest sounding level, whether that lowest sounding
level was at the surface or below the surface.  I corrected that to make sure
that the declared-height sounding code and the declared-pressure sounding code
are consistent with each other.






git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4193 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3669020ea53f1c9e6b31934db9a2c2e763662c6b