I added the "if ( l_hyper_dfsn ) then ..." statement to the portion of the statistical code (for both wp2 and wp3) that finalizes the implicit contribution of the 4th-order hyper-diffusion statistics.  This was done because of a false statistical computation that was taking place for wp2 statistical term wp2_4hd (4th-order hyper-diffusion).
In short, l_hyper_dfsn is set to .false. by default in the CLUBB code.
Thus, hyper-diffusion, as well as it's statistical contribution, are not
computed.  However, when hyper-diffusion is turned on, it's statistical
arrays are stored in zmscr13, zmscr14, zmscr15, zmscr16, and zmscr17.

Now, zmscr13, zmscr14, and zmscr15 are also used for the statistical
computations for wpxp in advance_xm_wpxp.  When hyper-diffusion is turned
on, those statistical arrays are overwritten in advance_wp2_wp3.  However,
since hyper-diffusion is not turned on, those arrays are not overwritten.

The problem occurred because the portion of the code that finalizes the
implicit contribution for the stat wp2_4hd did not contain a "l_hyper_dfsn"
flag.  Thus, even though hyper-diffusion is turned off, a false wp2_4hd
stat was being generated because of the left-over statistical garbage
from advance_xm_wpxp.

This code change fixes that bug, and the wp2 budget now balances again,
even when 4th-order hyper-diffusion is turned off.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4222 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ed312e00986e8ab7ed917ea1c1820df4c217acf8