I changed the code so that wp2_zt, which is the interpolated value of wp2, is output from advance_wp2_wp3.
In the current version of the code, wp2 is interpolated to find wp2_zt at the
beginning of subroutine advance_clubb_core.  This is sufficient for use everywhere.
However, within subroutine advance_wp2_wp3, wp2 needs to be interpolated again to
find wp2_zt, since wp2_zt is used in skewness clipping.  However, since the new
value of wp2_zt isn't needed again in advance_clubb_core, it is simply not output.

However, we are considered reorganizing the way that our advance_... subroutines
are ordered.  Now, wp2_zt is an important input variable for subroutine
advance_xp2_xpyp, where it is used to calculate the turbulent advection terms for
rtp2, thlp2, rtpthlp, up2, vp2, sclrp2, sclrprtp, and sclrpthlp.  It is also an
important variable for advance_xm_wpxp, where it is used to calculate the turbulent
advection terms for wprtp, wpthlp, and wpsclrp.  If the call to either subroutine
advance_xm_wpxp or subroutine advance_xp2_xpyp are moved to a point in the code
that is after the call to subroutine advance_wp2_wp3, the updated wp2_zt will
definitely be needed.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4226 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c3cfdfa73fe6b0c3aec1b14ad283cf9199bed6a3