I have done some reorganization of the order of subroutine calls in subroutine advance_clubb_core.
The goal was to move the call to advance_xp2_xpyp to a location in
advance_clubb_core that is by the other three advance_... subroutines.
An ordering of all four advance_... subroutines needed to be found
that gave the best possible results.

A couple of orderings were found that gave good results.  The one
being committed to the code, at least for right now, is:

advance_xm_wpxp
advance_xp2_xpyp
advance_wp2_wp3
advance_windm_edsclrm

Additionally, the call to sfc_var is now moved to a point later in
advance_clubb_core, which still immediately precedes the call to the
advance_... subroutines.  This allows for Lscale (or Lscale_up) to be
used as an input for the Andre et al. 78 sfc variance scheme.

Another benefit is that sigma_sqd_w, tau_zm, Kh_zt, and wpthvp
do not have to be saved from timestep-to-timestep anymore.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4228 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3d2758e4a5457e0739a0eb6fc0d4950d0d0bc7cd