The file advance_xp2_xpyp_module.F90 has been been coded for anelastic turbulent advection.
The budgets still balance, and when the equations are "reduced to Boussinesq",
we still get the old results back, except in very highly sensitive cases.
After all, the discretization is technically different.  This is exactly what
we saw after the changes were made to wp2 and wp3.

With the exception of um and vm (and edsclrm), all of the CLUBB equations
have had their turbulent advection terms (as well as turbulent production
for wp3) coded for an anelastic equation set.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4291 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bf9b1ea2057f1b9cd6fb47896ee563d27b9cd929