I am officially committing the anelastic changes for advance_windm_edsclrm_module.F90.  When the code is reduced to Boussinesq form, all the new results exactly match the old results.  So far, the overall anelastic results look good.
I am also committing changes to the description portion of
the monotonic flux limiter.  Even though xm and wpxp have
already been changed to anelastic format, the monotonic flux
limiter has not been updated to match that change.  This is
the first step, which is altering the description to describe
the anelastic form of the monotonic flux limiter.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4295 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 9b0f3b532640fb82d6b5aebe130ba57cd01cfb9a