The subroutine inverse_hydrostatic has been altered.  There are multiple points to make about this change.
The basic idea of inverse_hydrostatic is to determine the altitude of
input sounding levels that are given in terms of pressure coordinates.

The equation being solved for is:

d(exner)/dz = grav / ( Cp * thvm ).

The idea is to integrate and solve for z, such that:

- ( Cp / grav ) INT(exner_1:exner_2) thvm d(exner) = INT(z_1:z_2) dz;

and therefore:

z_2 = z_1 - ( Cp / grav ) INT(exner_1:exner_2) thvm d(exner).

The previous version of the code assumed thvm to be constant over a
layer that extended halfway between the sounding levels on either side.

The new version of the code considers thvm to vary linearly (with
respect to exner) between successive sounding levels.  Therefore, if
the value of thvm is the same at two successive sounding levels, then
it will be constant over that layer.  I have added new code to make
these calculations.  The new code has been tested, and it works well.



Along the way, it was discovered that the previous version of the code
was inaccurate.  Here is an example:  consider a pressure sounding over
it's lowest three levels -- level 3 has a pressure of about 931.73 mb,
level 2 has a pressure of about 965.43, and level 1 has a pressure of
exactly 1000 mb.  These values are chosen because they make the exner
values simple:  0.98, 0.99, and 1.0 for levels 3, 2, and 1, respectively.
Furthermore, consider thvm to be exactly 300 K at each one of these
levels.  Thus, thvm will be considered to be a constant of 300 K for the
duration of this example.  In the case of a constant thvm, the equation
for z can be expressed very simply as:

z_2 = z_1 - ( Cp / grav ) * thvm * ( exner_2 - exner_1 ).

Furthermore, consider the lowest sounding level to be right at the ground,
with a surface pressure of 1000 mb. and an initial height of 0 m.

Between levels 1 and 2 (and between levels 2 and 3), the difference in
exner is exactly -0.01.  When multiplied by a constant 300 K, and
furthermore by a constant ( Cp / grav ), the altitude of level 2 is found
to be approximately 307.24 m., and the altitude of level 3 is found to be
approximately 614.48 m.

The previous code interpolated exner to some hypothetical exner_zm between
the sounding levels (and considered thvm to be constant over that layer),
such that exner_zm(1) would be .995 and exner_zm(2) would be .985 in the
above example.  Now, the previous code did come up with 307.24 m. between
successive exner_zm sounding levels for this hypothetical test case.
However, it then performed an interpolation back to sounding levels.  In
the process of doing, so it set z(1) to 0 and interpolated z(2) as being
"halfway" between successive zm_snd levels.  This resulted in a value of
z(2) that was 153.62 m., or exactly half of what it should have been.



Along the way, I discovered further complications with the previous version
of the code.  I looked at the sounding for TWP-ICE, which has a pressure
at it's sounding lowest level of 1015 mb., yet had a surface pressure of
1004.4 mb.  Therefore, the lowest sounding level is not at the surface.
Nowhere did inverse_hydrostatic account for this.  I could determine the
depth of the layers between successive sounding layers, but I could not
determine where the surface was, and therefore could not determine altitude
in an absolute (above surface level) sense.

To solve this problem, I brought in psfc (which is at altitude zm_init) and used
it to help find above-ground altitude.  Simply, psfc is converted to exner_sfc.
Then, exner_sfc is compared to the sounding profile of exner.  The altitude
difference between exner_sfc and exner(1) (which is exner at sounding level 1) is
computed.  Now that the difference between the surface altitude and then altitude
of the bottom of the sounding is known, the altitude of the sounding profile can
be adjusted accordingly.

I also discovered that an obscure, poorly documented line in input_interpret was
overwriting the lowest sounding level pressure with surface pressure.  I got rid
of this.

I also noticed that the surface pressures given for all cloud feedback cases did
not match their lowest sounding level pressures.  However, unlike TWP-ICE where
it's surface pressure is between two sounding levels, the surface pressures for
all cloud feedback cases is higher than the lowest sounding levels.  Thus, the
bottom of the sounding is above the surface.  My new code does account for this.
However, depending on the case, the lowest sounding levels were calculated to be
between 100 - 200 m. above the surface.  The lowest thermodynamic grid level is
at 25 m. for these cases.  Thus, read_sounding would not allow the code to continue
do to insufficient information.

Thus, in order to allow the code to run, and furthermore to match the new version
of cloud feedback to the old one (where psfc overwrote pressure at sounding level 1),
I am going to alter the values of pressure at sounding level 1 in the appropriate
sounding.in files to be equal to psfc.  Again, this is simply changing sounding.in
files to account for what we used to do discretely in the code.


That being said, I ran all cases to a successful completion.  This change effects
only the cloud feedback cases and TWP-ICE.

I also ran timestep_tests in order to test stability at longer timesteps.  The
results are exactly the same as with the previous version, except that TWP-ICE
now mysteriously runs at a 30-minute timestep, resulting in one less blow-up
with this version.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4316 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fbdeafb1bb86237302a91d10a0a3ed429833eabf