I am alterating the pressure at the lowest level of the input sounding to match code that formerly performed an overwrite in subroutine read_z_profile (input_interpret.F90).
Prior to this change, the value of pressure at the lowest sounding level
for each of the cloud feedback cases was 100000 Pa., which matches the
description on the website.  However, CLUBB previously overwrote this
pressure with the surface pressure (psfc) in subroutine read_z_profile.
The line of code that did that was obscure and not well-documented.  As
part of the transformation of the code surrounding subroutines hydrostatic
and inverse_hydrostatic, this line of code has been removed.  If the
user wants to set the pressure at the lowest sounding level to the surface
pressure, he/she will have to explicitly do so in the input sounding.
This is much better coding practice.

This will make the cloud feedback initial profiles basically the same
as they were before.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4318 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 10de458c8f7818015f545bd909b879303557324f