I am committing code to allow for a better calculation of initial dry density.
The first run through hydrostatic is now based on dry surface pressure.  This
allows for a better calculation of dry pressure, exner, and density at all
vertical levels.  The calculations are then adjusted to take into account
vapor pressure.  The results run_scm_all.bash show that there aren't any
changes to any of the cases (for the current Boussinesq form of the equations).



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4321 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6d85dc70307aee40b49093ddf877b329dd60db4b