I changed the code in subroutine initialize_clubb (clubb_driver.F90) for the case where the initial sounding is declared in terms of pressure coordinates.  The code now outputs not only total pressure, exner, and density, but also dry densities, which are then used to set rho_ds_zt and rho_ds_zm.
Along the way, I discovered a bug in the previous version of the code.  For cases where
pressure soundings are used and the temp/theta/theta_l variable is either temperature or theta,
initial cloud water mixing ratio was never computed.  Thus, the when the conversion to theta_l
was made, rcm could not have been taken into account properly.  Thus, the cases were given a
faulty initial thlm profile (as long as at least one of the grid levels was initially
saturated).

In the current code, all three cloud feedback cases and TWP-ICE use pressure soundings with
temperature (cloud feedbacks) or theta (TWP-ICE) as the temperature variable.  Thus, this bug
had the potential to effect all of those cases.  It looks like cloud feedback s6 and TWP-ICE
are effected, for their profiles change on a plotgen.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4322 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6a92d7ba6619ae22f404dfe16d8dea701cfe3ad5