I changed a few comments in subroutine initialize_clubb_variables in order to better describe theta_d.
Just to make some notes ... I used print statements to look at the
results of the dry code versus the moist code for the density variables.

I printed the output from the dry code, which is rho_dry (and rho_dry_zm),
and I also printed the value of rtm (and zt2zm( rtm )).  Since

rho = rho_dry * ( 1.0 + rtm ),

I decided to print out the value of that expression.  I then compared it
to the printed value of rho (and rho_zm) as determined by the model code.

I found that there was very high accuracy, especially for cases where
the sounding is given in pressure coordinates (and thus integration is
not needed to determine exner/pressure, rather pressure is given).
There was at least three decimal place accuracy at every vertical level.
For cases were the sounding is given in altitude coordinates (and
pressure/exner is found through integration), the results were just as
accurate (to at least three decimal places) in the lower parts of the
vertical profile.  The results did decline in accuracy (comparing rho
as calculated above to rho obtained from the code) in the upper parts
of most runs (that use an altitude sounding).  However, these differences
were only on the order of about 0.01 kg/m^3^.  Of course, in cases that go
way, way up in altitude, the difference grows smaller and smaller again as
the values of the densities get to be very small.  Any small differences
can be chalked up to the calculation of pressure/exner or the fact that
an approximation is used to obtain theta_v for the integration of the moist
code.  Overall, the results are a great success!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4339 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 69085f5bb914d19fe45dfe60edb9c0990b0874f1