I set up the buoyancy production (and related pressure terms) for the anelastic set of equations.
Very generally, the previous (Boussinesq) form of the equations used:

( grav / T0 ) * xpthvp;

where T0 is a constant reference temperature, and "xp" in xpthvp depends
on the equation it is being used in -- for example "xp" is "wp2" for the
wp3 equation, "thlp" for the wpthlp equation, "rtp" for the wprtp equation,
and "wp" for the wp2, up2, and vp2 equations.

For the anelastic code, the above equation becomes:

( grav / thv_ds ) * xpthvp;

where thv_ds is the dry, (static) base-state theta_v calculated from the
initial profile, which varies only with height.  Nonetheless, it is usually
not constant over the entire vertical profile.

Since the Boussinesq form of the equations is still in play at this point,
thv_ds_zt and thv_ds_zm are overwritten by T0, and the new model results
exactly match the old model results.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4342 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: cf285f03cd8e4a48c2d39632640af451dc67d411