After further review ... the ruling on the field has been reversed ... first down!
This commitment alters the way in which the dry variables are being calculated
for the case where the sounding is given in altitude.

In previous revisions (see r4321, r4331), I changed the first call to subroutine
hydrostatic so that it was completely dry --- the inputs were dry surface pressure
and dry theta, and the outputs where dry pressure, exner, and density.  For
purposes of calculating rcm (as well as thlm/thm), approximate total pressure was
calculated from dry pressure and rtm, and then total exner was calculated from
total pressure.  Then, thvm was calculated, and hydrostatic was called again to
provide more accurate values of total pressure, exner, and density.

With this revision, I am reverting back to the old method of using total surface
pressure and theta/theta_l as a proxy for theta_v for the inputs to subroutine
hydrostatic.  The outputs are approximate total pressure, exner, and density.
I then use these variables directly to calculate rcm (as well as thm/thlm).  Then,
thvm is calculated from thm, rtm, and rcm.  After that, subroutine hydrostatic is
called to provide a more accurate value of total pressure, exner, and density.
So far, what I have listed is the old method.  After this, the dry values of the
variables are calculated.

This was done for a variety of reasons:

1)  This parallelizes the altitude sounding code with the pressure sounding code,
which already uses a method in this order (dry variables calculated after "total"
variables calculated), since the pressure sounding already gives total pressure
at each level.

2)  Tests were conducted to compare rho_dry, a calculation of total rho based
on rho_dry and rtm, and the actual total rho.  Ideally, the calculation of total
rho and the actual value of total rho should be nearly identical.  Keep in mind
that rho_dry and rho are not calculated from each other in the code, but rather
are determined by the equation of state.  Anyway, the two values had always been
in good agreement for the cases that use a pressure sounding.  However, there
was a deviation that increased with height for cases that use an altitude sounding.
This must have been due to the call to hydrostatic using dry variables.  The
dry integration threw things off a little bit.  Anyway, the values should now
be in very good agreement for cases that use an altitude sounding.

3)  After I made the original change, LBA developed more of an anvil top and
contained cloud near the beginning timesteps.  A subsequent change reduced the
amount of cloudiness in the early LBS timesteps, but as it turns out, there still
was some.  I don't know the reasons why this happened, but this change reverts
LBA back to its old form.


The only cases visibly changed are LBA, as described in point #3, and June 25
(very slightly), which had also changed very slightly when the original change
was made.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4354 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 5a4b2aa9bbf48ae90629513a2a1bf3a2f355fc82