I have updated subroutine read_z_profile in input_interpret.F90.  This change is in regards to the code that reads in a sounding given on pressure levels and calculates the altitude of the sounding levels.
I have changed the calculation of theta_v so that it uses the exact,
non-linear version (the same version that is now used in the CLUBB
inititalization -- see CLUBB r4352).  I think that accuracy is more
important here because this code determines the height of the sounding
levels, which then are used to set the initial values of the main
model variables at every model vertical level.

In setting up for this change, I changed the code slightly so that
a sounding given in temperature would convert to theta (rather than
convert to theta_l), a sounding given in theta would remain as theta
(rather than convert to theta_l), and a sounding given in theta_l
would convert to theta (rather than remain as theta_l).  This was
done because the version of the equation for theta_v now in use
uses theta rather than theta_l.

A plotgen showed no differences between this version of the code
and the prior version of the code.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4355 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 83b7919304d045582e36bb297e5d261f3bb1819a