I made a slight alteration to the initialization code, in order to help acheive the best results for cases that are initially saturated.
Rather than simply send in thm/thlm into subroutine hydrostatic for
the first call to subroutine hydrostatic, an approximate thvm
is now computed based on thm/thlm and rtm.  This is done because
the equation used in subroutine hydrostatic actually calculates
exner, pressure, and density based on thvm, not thm or thlm.
Of course, a more accurate thvm is calculated later, after rcm
is computed.  Based on that more accurated value of thvm, a more
accurate pressure, exner, and density are calculated in the second
call to subroutine hydrostatic.

However, based on problems encountered in the past where a faulty
initially saturated profile caused problems for the LBA case, I 
decided to produce the best exner/pressure possible from the first
call to subroutine hydrostatic.

Considering there to be some moisture in the initial sounding:
theta_v > theta >= theta_l.  Therefore:

(1/theta_l) >= (1/theta) > (1/theta_v).

Now, the form of the hydrostatic equation used in subroutine
hydrostatic is:

d(exner)/dz = - ( grav / Cp ) * ( 1 / theta_v ).

Using theta or theta_l instead of theta_v produces a value of
d(exner)/dz that is too large in magnitude.  Thus, exner falls
of too much with height, and therefore pressure falls off too
much with height.  Now, when the formula for saturation mixing
ratio is called, temperature is calculated from theta*exner.
A value of exner that is too small produces a temperture that is
too small.  Therefore, saturation mixing ratio is artificially
too small.  This may cause saturation too occur when it shouldn't,
or too much saturation to occur if the profile is supposed to
be saturated.

This fix should help alleviate any problem of the aforementioned
nature.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4372 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7cd79cd97d297803a357291f37021e559ba3269a