I did a little bit of housekeeping on the code that sets the upper boundary conditions in advance_xp2_xpyp_module.F90 and in advance_wp2_wp3_mod.F90.  However, there aren't any changes to the actual code itself.
I removed the comments that stated that the fixed value of a variable
at the upper boundary was changed from 0 to some threshold because
fill_holes would otherwise fill the hole at that top level.  The reason
I removed those comments is because they are not true.  In fact, they
never were true.  I have always had a catch in fill_holes.F90 that
ensures that, for a momentum-level variable, the uppermost and lowermost
vertical levels (gr%nnzp and 1, respectively) are NOT considered
in regards to hole-filling.  This was done so that the fixed values
of those variables at the boundaries remain just that, fixed values.

I also removed a little bit of the old junk lying around.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4374 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 26f95bca9a67d8035bcb4f72025db432339c362e