The code used by the local version of the KK evaporation equation to directly compare the local version to the upscaled version by calculating Supersaturation, S, as a function of extended liquid water mixing ratio, s, needed an adjustment.
Specifically, r_s was written as a function of temperature and
pressure (which it usually is), such that:  r_s = rsat(T,p).  However,
in this case, it needed to be a function of liquid water temperature
and pressure, such that:  r_sl = rsat(T_l,p).

Likewise, Beta(T_l) needed to be a function of liquid water temperature.
It had been coded as a function of temperature, Beta(T).

This is now corrected.  Both the local and the upscaled equation sets
now use the same form of these terms.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4433 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7449e0496f704c3dd50b920e40dbe116396e69a0