I am making a few changes to the KK microphysics file.  These changes mainly correspond to the writing of G(T,p) as G(T_l,p) in the new version of Part I of the paper.
Unlike the expressions for r_sl ( r_s(T_l,p) ) and Beta_Tl ( Beta(T_l) ),
expressing G(T,p) as G(T_l,p) is NOT required in order to link
Supersatuation, S, to extended liquid water mixing ratio, s.  However,
since all other "T" variables relating to the coefficient terms in the
KK evaporation equation are expressed in terms of T_l (due to being
necessary for the approxmations linking S to s, as mentioned above), it
is easier (for paper writing purposes) to include G(T_l,p) in that group
and then have one "coefficient" variable that is simple a function of T_l.

Thus, since the paper is written that way, the code must be written that
way too.

While the effects of this change are very small, the nice thing is that
it does slightly decrease evaporation, thereby slightly increasing rain
water at the surface, which is what we need!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4436 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 83dc8b899995ccc332b4e70b525ba1566e8008e9