Missing variables will no longer cause disappearance of plots in the form of broken images. Instead, missing variables are replaced with a 0. This is implemented in the form of try / catch blocks.
To support the try / catch blocks, VariableReadGrADS.m needed to be
modified to throw a 'GrADS variable not found!' error if a variable
is not found.

The NetCDF functions already supported an equivalent to the above
error.

This has been tested and works with NetCDF and GrADS profile plots.

Timeseries functionality is currently handicapped and will only plot
a line of 0s if any variable is missing.

This is in relation to tickets #214 and #271.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4476 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ae8a4942f19acbecdc8a42abeac3ab13f8a36ebe