Removed now unused DetermineCloudVar.m
Added 'CaseFunctions' directory for functions intended for use in case
files. Any functions added to this directory are automatically made
available for use in case files.

Added UseFirstArgIfExist function. This function has two arguments,
'default' and 'fallback.' If default exists, it is used, otherwise the
data contained in fallback is used. This works because of the fact
variables which do not exist are assigned a single zero, so if a 
variable has a size of one, if does not exist. This function will
not work for cases which have a single vertical level.

PlotCreator was modified to allow for the CaseFunctions directory to
work as described.

This is all related to ticket #314.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4490 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: bde931b89112ef2e808a0468bb5359520922e2c6