Created a new run script: run_scm_collect_suite.bash
This script runs a selected set of cases a few times each, to provide a standard timing test suite that can be used for better comparison of timing in the future.

In addition, the run_scripts README has been updated with information on the new script. Also, a note has been added to the README that when performing run_scm_collect(_suite) runs, the amount of other programs running on that computer should be minimized to reduce the inconsistency in the timing results.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4655 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e1c706719869171347df5161ef45a23181c7e030