For ticket #352.
Updated time_scm_all.bash to run only cases that do not use microphysics or radiation.

In addition, the script has been enhanced to aggregate the results of multiple runs--this is important, as the removal of many of the cases in CLUBB greatly reduces the runtime for a single "run" of run_scm_all, which means that more runs will need to be performed to get an overall runtime long enough to be useful for timing comparisons.

The script has a new required argument, which specifies the number of runs to perform. It can still take the path of an output file as an optional argument.

Lastly, the README has been updated to reflect the changes in the script.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@4821 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: d82946d2a81d35dbd46048c574eca0f0e001667d