For ticket #376.  This changeset contains the upwind differencing approximation of sedimenation (currently disabled), and modifies the sedimentation velocity clipping code to occur on thermodynamic levels and then interpolate to the momentum levels for the centered difference code.  For the Khairoutdinov Kogan microphysics the ordering is:
1. Compute mean volume radiation on thermo. grid;
2. Compute sedimentation velocity once on thermo. grid for the analytic code,
or n_micro_calls times on thermo. grid for latin hypercube sampling.
3. Clip sedimentation velocities on the thermo. grid (this is the change).
4. Use the thermo. level sedimentation velocities for upwind differencing or 
interpolate the sedimenation velocities to the momentum levels for the centered
differencing approximation.

In the old code the interpolation happened prior to clipping, so if a large
magnitude in velocity occured next to small magnitude in velocity this will
change the clipping.  Since this almost never happens unless the latin hypercube
sampling is enabled it only appears to affect the values in the rico_LH case.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5002 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c209c61ea238cda6dad0aec80734aaa55a62e79c